Current Monitor
Had some difficulties initially with the current monitor... Initially
assumed it was in the coding, shortly after, learned some things and
realized it was hardware.  As of now, the current monitor works better
than hoped for.  Will need to opto-couple to output to the micro
controller because the chip does not have built-in ground isolation.