House Cleaning
Did some organizing... added a few new functions and shifted some things
around.  Prepping the program to begin the PID controller.  Tested the
h-bridge with new PWM.  Works Great!  Realized time wasted in the
rpm_MOTOR calculation with the FOR loop.  Will probably make a 'faster'
circular array calculation not using the FOR.